+++ frontend/app/app/libs/axios/auth/__tests__/getAuthSessions.test.ts
@@ -1,7 +1,8 @@
 import { validUserData } from "@/__tests__/fixtures/auth/validUserData";
 import { mockApi, mockAxios, mockGet } from "@/__tests__/mocks/axios/api";
 import { mockAddAuthInfoToRequest } from "@/__tests__/mocks/cookie/mockLoadAuthInfo";
-import { getAuthSessions } from "@/app/libs/axios/auth/getAuthSessions";
+
+import { getAuthSessions } from "../getAuthSessions";
 
 import type { GetAuthSessionsData } from "@/app/types/axios/auth/getAuthSessions";
