+++ frontend/app/app/libs/axios/auth/__tests__/postAuthGuestSignIn.test.ts
@@ -1,9 +1,9 @@
-import { validSignInData } from "@/__tests__/fixtures/auth/validSignInData";
 import { mockUnexpectedResponse } from "@/__tests__/fixtures/unexpectedResponseData";
 import { mockApi, mockAxios, mockPost } from "@/__tests__/mocks/axios/api";
 import { UNEXPECTED_ERROR_MESSAGE } from "@/app/constants/errors/Messages";
-import { postAuthSignIn } from "@/app/libs/axios/auth/postAuthSignIn";
 import { saveAuthInfoFromHeader } from "@/app/libs/cookie/saveAuthInfo";
+
+import { postAuthGuestSignIn } from "../postAuthGuestSignIn";
 
 import type {
   PostAuthSignInErrorData,
@@ -13,7 +13,7 @@ import type {
 jest.mock("@/app/libs/cookie/saveAuthInfo");
 jest.mock("@/app/libs/axios/api", () => mockApi);
 
-describe("postAuthSignIn", () => {
+describe("postAuthGuestSignIn", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
@@ -23,55 +23,54 @@ describe("postAuthSignIn", () => {
   });
 
   describe("リクエストに成功した時", () => {
-    const mockPostAuthSignInSuccessData: PostAuthSignInSuccessData = {
+    const mockPostAuthGuestSignInSuccessData: PostAuthSignInSuccessData = {
       success: true,
       message: "success"
     };
 
-    const mockPostAuthSignInSuccessResponse = {
+    const mockPostAuthGuestSignInSuccessResponse = {
       headers: {
         "access-token": "mockToken",
         client: "mockClient",
         uid: "mockUid"
       },
-      data: mockPostAuthSignInSuccessData
+      data: mockPostAuthGuestSignInSuccessData
     };
 
     it("エラーをスローせず,saveAuthInfoHeaderが呼び出される", async () => {
-      mockPost.mockResolvedValue(mockPostAuthSignInSuccessResponse);
-      await expect(postAuthSignIn(validSignInData)).resolves.not.toThrow();
-      await postAuthSignIn(validSignInData);
+      mockPost.mockResolvedValue(mockPostAuthGuestSignInSuccessResponse);
+      await expect(postAuthGuestSignIn()).resolves.not.toThrow();
+      await postAuthGuestSignIn();
       expect(saveAuthInfoFromHeader).toHaveBeenCalledWith(
-        mockPostAuthSignInSuccessResponse
+        mockPostAuthGuestSignInSuccessResponse
       );
     });
   });
+
   describe("リクエストに失敗した時", () => {
     describe("エラーがAxiosから発生した場合", () => {
       describe("レスポンスの型がPostAuthSignInErrorDataを満たす場合", () => {
         it("適切なエラーメッセージがスローされる", async () => {
           const errorMessage = "Error message";
-          const mockPostAuthSignInErrorData: PostAuthSignInErrorData = {
+          const mockPostAuthGuestSignInErrorData: PostAuthSignInErrorData = {
             success: false,
             errors: [errorMessage]
           };
 
-          const mockPostAuthSignInErrorResponse = {
-            response: { data: mockPostAuthSignInErrorData }
+          const mockPostAuthGuestSignInErrorResponse = {
+            response: { data: mockPostAuthGuestSignInErrorData }
           };
 
           mockAxios.isAxiosError.mockReturnValue(true);
-          mockPost.mockRejectedValue(mockPostAuthSignInErrorResponse);
-          await expect(postAuthSignIn(validSignInData)).rejects.toThrow(errorMessage);
+          mockPost.mockRejectedValue(mockPostAuthGuestSignInErrorResponse);
+          await expect(postAuthGuestSignIn()).rejects.toThrow(errorMessage);
         });
       });
       describe("レスポンスの型がPostAuthSignInErrorDataを満たさない場合", () => {
         it("適切なエラーがスローされる", async () => {
           mockAxios.isAxiosError.mockReturnValue(true);
           mockPost.mockRejectedValue(mockUnexpectedResponse);
-          await expect(postAuthSignIn(validSignInData)).rejects.toThrow(
-            UNEXPECTED_ERROR_MESSAGE
-          );
+          await expect(postAuthGuestSignIn()).rejects.toThrow(UNEXPECTED_ERROR_MESSAGE);
         });
       });
     });
@@ -82,7 +81,7 @@ describe("postAuthSignIn", () => {
         const mockError = new Error(networkError);
         mockAxios.isAxiosError.mockReturnValue(false);
         mockPost.mockRejectedValue(mockError);
-        await expect(postAuthSignIn(validSignInData)).rejects.toThrow(networkError);
+        await expect(postAuthGuestSignIn()).rejects.toThrow(networkError);
       });
     });
   });
