+++ frontend/app/app/libs/axios/auth/__tests__/postAuthSignIn.test.ts
@@ -1,8 +1,10 @@
+import { validSignInData } from "@/__tests__/fixtures/auth/validSignInData";
 import { mockUnexpectedResponse } from "@/__tests__/fixtures/unexpectedResponseData";
 import { mockApi, mockAxios, mockPost } from "@/__tests__/mocks/axios/api";
 import { UNEXPECTED_ERROR_MESSAGE } from "@/app/constants/errors/Messages";
-import { postAuthGuestSignIn } from "@/app/libs/axios/auth/postAuthGuestSignIn";
 import { saveAuthInfoFromHeader } from "@/app/libs/cookie/saveAuthInfo";
+
+import { postAuthSignIn } from "../postAuthSignIn";
 
 import type {
   PostAuthSignInErrorData,
@@ -12,7 +14,7 @@ import type {
 jest.mock("@/app/libs/cookie/saveAuthInfo");
 jest.mock("@/app/libs/axios/api", () => mockApi);
 
-describe("postAuthGuestSignIn", () => {
+describe("postAuthSignIn", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
@@ -22,54 +24,55 @@ describe("postAuthGuestSignIn", () => {
   });
 
   describe("リクエストに成功した時", () => {
-    const mockPostAuthGuestSignInSuccessData: PostAuthSignInSuccessData = {
+    const mockPostAuthSignInSuccessData: PostAuthSignInSuccessData = {
       success: true,
       message: "success"
     };
 
-    const mockPostAuthGuestSignInSuccessResponse = {
+    const mockPostAuthSignInSuccessResponse = {
       headers: {
         "access-token": "mockToken",
         client: "mockClient",
         uid: "mockUid"
       },
-      data: mockPostAuthGuestSignInSuccessData
+      data: mockPostAuthSignInSuccessData
     };
 
     it("エラーをスローせず,saveAuthInfoHeaderが呼び出される", async () => {
-      mockPost.mockResolvedValue(mockPostAuthGuestSignInSuccessResponse);
-      await expect(postAuthGuestSignIn()).resolves.not.toThrow();
-      await postAuthGuestSignIn();
+      mockPost.mockResolvedValue(mockPostAuthSignInSuccessResponse);
+      await expect(postAuthSignIn(validSignInData)).resolves.not.toThrow();
+      await postAuthSignIn(validSignInData);
       expect(saveAuthInfoFromHeader).toHaveBeenCalledWith(
-        mockPostAuthGuestSignInSuccessResponse
+        mockPostAuthSignInSuccessResponse
       );
     });
   });
-
   describe("リクエストに失敗した時", () => {
     describe("エラーがAxiosから発生した場合", () => {
       describe("レスポンスの型がPostAuthSignInErrorDataを満たす場合", () => {
         it("適切なエラーメッセージがスローされる", async () => {
           const errorMessage = "Error message";
-          const mockPostAuthGuestSignInErrorData: PostAuthSignInErrorData = {
+          const mockPostAuthSignInErrorData: PostAuthSignInErrorData = {
             success: false,
             errors: [errorMessage]
           };
 
-          const mockPostAuthGuestSignInErrorResponse = {
-            response: { data: mockPostAuthGuestSignInErrorData }
+          const mockPostAuthSignInErrorResponse = {
+            response: { data: mockPostAuthSignInErrorData }
           };
 
           mockAxios.isAxiosError.mockReturnValue(true);
-          mockPost.mockRejectedValue(mockPostAuthGuestSignInErrorResponse);
-          await expect(postAuthGuestSignIn()).rejects.toThrow(errorMessage);
+          mockPost.mockRejectedValue(mockPostAuthSignInErrorResponse);
+          await expect(postAuthSignIn(validSignInData)).rejects.toThrow(errorMessage);
         });
       });
       describe("レスポンスの型がPostAuthSignInErrorDataを満たさない場合", () => {
         it("適切なエラーがスローされる", async () => {
           mockAxios.isAxiosError.mockReturnValue(true);
           mockPost.mockRejectedValue(mockUnexpectedResponse);
-          await expect(postAuthGuestSignIn()).rejects.toThrow(UNEXPECTED_ERROR_MESSAGE);
+          await expect(postAuthSignIn(validSignInData)).rejects.toThrow(
+            UNEXPECTED_ERROR_MESSAGE
+          );
         });
       });
     });
@@ -80,7 +83,7 @@ describe("postAuthGuestSignIn", () => {
         const mockError = new Error(networkError);
         mockAxios.isAxiosError.mockReturnValue(false);
         mockPost.mockRejectedValue(mockError);
-        await expect(postAuthGuestSignIn()).rejects.toThrow(networkError);
+        await expect(postAuthSignIn(validSignInData)).rejects.toThrow(networkError);
       });
     });
   });
